+++ front/src/applications/operationalStudies/views/Scenario/components/ImportTimetableItem/helpers/parseGraouTrains.ts
@@ -7,6 +7,38 @@ import { Duration } from 'utils/duration';
 import findValidTrainNameKey from './findValidTrainNameKey';
 import rollingstockOpenData2OSRD from '../rollingstock_opendata2osrd.json';
 
+// The Graou API returns only part of the UIC for French stations: the last
+// digit is missing. It's a checksum which uses the Luhn algorithm.
+export const populateUicChecksumInStep = (uic: string): string => {
+  if (!uic.startsWith('87') || uic.length >= 8) {
+    return uic;
+  }
+
+  const rawCi = uic.substring(2);
+  let checksum = 0;
+  for (let i = 0; i < rawCi.length; i++) {
+    const digit = Number(rawCi[i]);
+    if ((i + 1) % 2 === (rawCi.length + 1) % 2) {
+      checksum += digit;
+    } else if (digit > 4) {
+      checksum += 2 * digit - 9;
+    } else {
+      checksum += 2 * digit;
+    }
+  }
+  checksum = (10 - (checksum % 10)) % 10;
+
+  return uic + String(checksum);
+};
+
+export const populateUicChecksumInTrainSchedules = (trainSchedules: GraouTrainSchedule[]): void => {
+  for (const trainSchedule of trainSchedules) {
+    for (const step of trainSchedule.steps) {
+      step.uic = populateUicChecksumInStep(step.uic);
+    }
+  }
+};
+
 export function filterInvalidSteps(importedTrainSchedules: GraouTrainSchedule[]): {
   filteredTrains: GraouTrainSchedule[];
   modifiedTrainsNumbers: string[];
